+++ wezterm-mux-server-impl/src/dispatch.rs
@@ -176,6 +176,18 @@ where
                     .await?;
                 stream.flush().await.context("flushing PDU to client")?;
             }
+            Ok(Item::Notif(MuxNotification::TabTitleChanged { tab_id, title })) => {
+                Pdu::TabTitleChanged(codec::TabTitleChanged { tab_id, title })
+                    .encode_async(&mut stream, 0)
+                    .await?;
+                stream.flush().await.context("flushing PDU to client")?;
+            }
+            Ok(Item::Notif(MuxNotification::WindowTitleChanged { window_id, title })) => {
+                Pdu::WindowTitleChanged(codec::WindowTitleChanged { window_id, title })
+                    .encode_async(&mut stream, 0)
+                    .await?;
+                stream.flush().await.context("flushing PDU to client")?;
+            }
             Ok(Item::Notif(MuxNotification::ActiveWorkspaceChanged(_))) => {}
             Ok(Item::Notif(MuxNotification::Empty)) => {}
             Err(err) => {